+++ x/gov/simulation/operations_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/cosmos/gogoproto/proto"
 	"github.com/stretchr/testify/require"
 
+	"cosmossdk.io/core/header"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
 
@@ -34,7 +35,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/gov/types"
 	v1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	"github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
-	_ "github.com/cosmos/cosmos-sdk/x/params"
 	_ "github.com/cosmos/cosmos-sdk/x/staking"
 	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
 )
@@ -213,7 +213,7 @@ func TestSimulateMsgCancelProposal(t *testing.T) {
 	suite, ctx := createTestSuite(t, false)
 	app := suite.App
 	blockTime := time.Now().UTC()
-	ctx = ctx.WithBlockTime(blockTime)
+	ctx = ctx.WithHeaderInfo(header.Info{Time: blockTime})
 
 	// setup 3 accounts
 	s := rand.NewSource(1)
@@ -225,7 +225,7 @@ func TestSimulateMsgCancelProposal(t *testing.T) {
 	contentMsg, err := v1.NewLegacyContent(content, suite.GovKeeper.GetGovernanceAccount(ctx).GetAddress().String())
 	require.NoError(t, err)
 
-	submitTime := ctx.BlockHeader().Time
+	submitTime := ctx.HeaderInfo().Time
 	params, _ := suite.GovKeeper.Params.Get(ctx)
 	depositPeriod := params.MaxDepositPeriod
 
@@ -262,7 +262,7 @@ func TestSimulateMsgDeposit(t *testing.T) {
 	suite, ctx := createTestSuite(t, false)
 	app := suite.App
 	blockTime := time.Now().UTC()
-	ctx = ctx.WithBlockTime(blockTime)
+	ctx = ctx.WithHeaderInfo(header.Info{Time: blockTime})
 
 	// setup 3 accounts
 	s := rand.NewSource(1)
@@ -274,7 +274,7 @@ func TestSimulateMsgDeposit(t *testing.T) {
 	contentMsg, err := v1.NewLegacyContent(content, suite.GovKeeper.GetGovernanceAccount(ctx).GetAddress().String())
 	require.NoError(t, err)
 
-	submitTime := ctx.BlockHeader().Time
+	submitTime := ctx.HeaderInfo().Time
 	params, _ := suite.GovKeeper.Params.Get(ctx)
 	depositPeriod := params.MaxDepositPeriod
 
@@ -312,7 +312,7 @@ func TestSimulateMsgVote(t *testing.T) {
 	suite, ctx := createTestSuite(t, false)
 	app := suite.App
 	blockTime := time.Now().UTC()
-	ctx = ctx.WithBlockTime(blockTime)
+	ctx = ctx.WithHeaderInfo(header.Info{Time: blockTime})
 
 	// setup 3 accounts
 	s := rand.NewSource(1)
@@ -324,7 +324,7 @@ func TestSimulateMsgVote(t *testing.T) {
 	contentMsg, err := v1.NewLegacyContent(v1beta1.NewTextProposal("Test", "description"), govAcc)
 	require.NoError(t, err)
 
-	submitTime := ctx.BlockHeader().Time
+	submitTime := ctx.HeaderInfo().Time
 	params, _ := suite.GovKeeper.Params.Get(ctx)
 	depositPeriod := params.MaxDepositPeriod
 
@@ -361,7 +361,7 @@ func TestSimulateMsgVoteWeighted(t *testing.T) {
 	suite, ctx := createTestSuite(t, false)
 	app := suite.App
 	blockTime := time.Now().UTC()
-	ctx = ctx.WithBlockTime(blockTime)
+	ctx = ctx.WithHeaderInfo(header.Info{Time: blockTime})
 
 	// setup 3 accounts
 	s := rand.NewSource(1)
@@ -372,7 +372,7 @@ func TestSimulateMsgVoteWeighted(t *testing.T) {
 	govAcc := suite.GovKeeper.GetGovernanceAccount(ctx).GetAddress().String()
 	contentMsg, err := v1.NewLegacyContent(v1beta1.NewTextProposal("Test", "description"), govAcc)
 	require.NoError(t, err)
-	submitTime := ctx.BlockHeader().Time
+	submitTime := ctx.HeaderInfo().Time
 	params, _ := suite.GovKeeper.Params.Get(ctx)
 	depositPeriod := params.MaxDepositPeriod
 
@@ -423,7 +423,6 @@ func createTestSuite(t *testing.T, isCheckTx bool) (suite, sdk.Context) {
 			configurator.NewAppConfig(
 				configurator.AuthModule(),
 				configurator.TxModule(),
-				configurator.ParamsModule(),
 				configurator.BankModule(),
 				configurator.StakingModule(),
 				configurator.ConsensusModule(),