+++ x/nft/simulation/genesis.go
@@ -3,6 +3,7 @@ package simulation
 import (
 	"math/rand"
 
+	"cosmossdk.io/core/address"
 	"cosmossdk.io/x/nft"
 
 	"github.com/cosmos/cosmos-sdk/types/module"
@@ -25,12 +26,16 @@ func genClasses(r *rand.Rand, accounts []simtypes.Account) []*nft.Class {
 }
 
 // genNFT returns a slice of nft.
-func genNFT(r *rand.Rand, classID string, accounts []simtypes.Account) []*nft.Entry {
+func genNFT(r *rand.Rand, classID string, accounts []simtypes.Account, ac address.Codec) []*nft.Entry {
 	entries := make([]*nft.Entry, len(accounts)-1)
 	for i := 0; i < len(accounts)-1; i++ {
 		owner := accounts[i]
+		oast, err := ac.BytesToString(owner.Address.Bytes())
+		if err != nil {
+			panic(err)
+		}
 		entries[i] = &nft.Entry{
-			Owner: owner.Address.String(),
+			Owner: oast,
 			Nfts: []*nft.NFT{
 				{
 					ClassId: classID,
@@ -44,7 +49,7 @@ func genNFT(r *rand.Rand, classID string, accounts []simtypes.Account) []*nft.En
 }
 
 // RandomizedGenState generates a random GenesisState for nft.
-func RandomizedGenState(simState *module.SimulationState) {
+func RandomizedGenState(simState *module.SimulationState, ac address.Codec) {
 	var classes []*nft.Class
 	simState.AppParams.GetOrGenerate(
 		"nft", &classes, simState.Rand,
@@ -56,7 +61,7 @@ func RandomizedGenState(simState *module.SimulationState) {
 		"nft", &entries, simState.Rand,
 		func(r *rand.Rand) {
 			class := classes[r.Int63n(int64(len(classes)))]
-			entries = genNFT(r, class.Id, simState.Accounts)
+			entries = genNFT(r, class.Id, simState.Accounts, ac)
 		},
 	)